// ui/src/generated/types/updateStorageInfoParams.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * Smartphone Test Farm
 * Control and manages real Smartphone devices from browser and restful apis
 * OpenAPI spec version: 2.4.3
 */

export type UpdateStorageInfoParams = {
  /**
   * Device place
   */
  place?: string
  /**
   * Device storageId
   */
  storageId?: string
  /**
   * Device adbPort
   */
  adbPort?: number
}